Sensor Monitoring System Service and Client
This project consists of a database, a web service which uses this constructed database and a client which uses this developed web service. Also, the clients pages are responsive. The graphical demonstration of sensor datas was generated by the help of Google Charts .js. This project is developed by me in my internship period.